+++ packages/core/src/index.ts
@@ -17,8 +17,9 @@ export {
 } from "./search/request.js";
 export { searchMarkdown, type SearchResponseV1 } from "./search/search.js";
 export { loadConfig } from "./config/load.js";
-export { isAbsoluteLocalPath } from "./config/locations.js";
+export { isAbsoluteLocalPath, resolveConfigPath } from "./config/locations.js";
 export { selectProject } from "./config/project.js";
+export { validateConfig } from "./config/validate.js";
 export type { ResolvedConfig, ResolvedProject } from "./config/types.js";
 export {
   AgentMarkdownError,